+++ src/algorithm/backTrack/generateParenthesis.ts
@@ -1,21 +1,21 @@
 // https://leetcode-cn.com/problems/generate-parentheses/
 // 22
 export function generateParenthesis(n: number): string[] {
-  return recursion([], '', 0, 0, n);
+  return dfs([], '', 0, 0, n);
 }
 
-function recursion(result: string[], path: string, left: number, right: number, n: number): string[] {
+function dfs(result: string[], path: string, left: number, right: number, n: number): string[] {
   if (left === n && right === n) {
     result.push(path);
     return result;
   }
 
   if (left < n) {
-    recursion(result, `${path}(`, left + 1, right, n);
+    dfs(result, `${path}(`, left + 1, right, n);
   }
 
   if (left > right) {
-    recursion(result, `${path})`, left, right + 1, n);
+    dfs(result, `${path})`, left, right + 1, n);
   }
 
   return result;