+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# (2021-12-17)
+# (2021-12-23)
 
 ### Bug Fixes
 
@@ -90,6 +90,7 @@
 - **algorithm/slidingWindow:** add lengthOfLongestSubstring ([50e7a5f](https://github.com/air-supply94/js-algorithm/commit/50e7a5f08be352c3201605a599f0ecf668f4228c))
 - **algorithm/slidingWindow:** add minWindow ([1316b41](https://github.com/air-supply94/js-algorithm/commit/1316b4109466ed38bea1dad44868e4cc84d9d4aa))
 - **algorithm/slidingWindow:** add numSubarrayProductLessThanK ([e94d152](https://github.com/air-supply94/js-algorithm/commit/e94d152f3d3412a624b388e6b1040c6c8c7f9304))
+- **algorithm/slidingWindow:** 增加区间说明 ([9e5168c](https://github.com/air-supply94/js-algorithm/commit/9e5168c90d7f960e74405a56c742884c44aa7d7d))
 - **array:** add groupStr ([d8e4365](https://github.com/air-supply94/js-algorithm/commit/d8e43659c0b2cc3ed5fdc229e8db82269fb08281))
 - **array:** add maxProfit ([601e32a](https://github.com/air-supply94/js-algorithm/commit/601e32ad487855c6359500e71078a5b4ed42ff65))
 - **array:** add multiplyExceptSelf ([545e5e9](https://github.com/air-supply94/js-algorithm/commit/545e5e9ec432554a9e120f3f40178c8078442b51))
@@ -140,6 +141,7 @@
 - **dataStructure:** add DisjointSet ([2903fb3](https://github.com/air-supply94/js-algorithm/commit/2903fb3bc1f1b9cee927aca551338cc2d3559efe))
 - **dataStructure:** add minCostConnectPoints ([79f255d](https://github.com/air-supply94/js-algorithm/commit/79f255d8e74741b5baea41072dfd723a3c47459b))
 - **dataStructure:** update disjointSet ([569f487](https://github.com/air-supply94/js-algorithm/commit/569f487b6a1b64a0091718eb588c672ae0dd2a25))
+- **dataStructure:** 迁移并查集、图 ([8a10304](https://github.com/air-supply94/js-algorithm/commit/8a103043b3a161517c394de2ff2f378b11666c96))
 - **dataStructure/bst:** add findMode ([aa0d969](https://github.com/air-supply94/js-algorithm/commit/aa0d969c0672a81c643ec9260096ec6b1920955a))
 - **dataStructure/bst:** update getHeight ([972e0c4](https://github.com/air-supply94/js-algorithm/commit/972e0c43279a8d26aee1d49e00cd827fa93a4e7a))
 - **dataStructure/design:** add FrequencyStack ([96b15a0](https://github.com/air-supply94/js-algorithm/commit/96b15a0b1ffa45e9824c079ad06b97a65ba7f3c0))